third_party/WebKit/LayoutTests/animations/css-animation-overrides-svg-presentation-attribute-animation-expected.txt

CONSOLE WARNING: Hyphenated naming in Web Animations keyframes is deprecated and will be removed in M51, around May 2016. Please use camelCase instead. See https://www.chromestatus.com/features/5650817352728576 for more details.
 This is a testharness.js-based test.
FAIL CSS animations always override SVG presentation attribute animations assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
